+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to **nimbiCMS** will be documented in this file.
 - ...
 - **[FEAT]** Added cosmetic URLs (`website/#/slug[#anchor][?params]`) on top of canonical URLs (`website/?page=slug[#anchor][?params]`)
 - **[FEAT]** Improved SEO management
-- **[FEAT]** Added a dynamic sitemap generator
+- **[FEAT]** Added a dynamic sitemap
 
 ## v1.0.5